requirements.txt
flask>=2.0.0
snowflake-connector-python>=3.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.21.0
requests>=2.25.0
werkzeug>=2.0.0
jinja2>=3.0.0
gunicorn>=20.0.0
